requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.1
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
blinker==1.7.0
celery==5.3.6
certifi==2023.11.17
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.1.3
django-celery-results==2.5.1
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
ecdsa==0.18.0
Flask==3.0.0
frozenlist==1.4.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jose==1.0.0
kombu==5.3.5
magic-filter==1.0.12
MarkupSafe==2.1.3
msgpack==1.0.7
multidict==6.0.4
prompt-toolkit==3.0.43
pyasn1==0.5.1
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
pyTelegramBotAPI==4.15.2
python-dateutil==2.8.2
python-jose==3.3.0
python-telegram-handler==2.2.1
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
rsa==4.9
sentry-sdk==1.39.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
whitenoise==6.6.0
yarl==1.9.4
